+++ Modules/Users/Providers/UsersServiceProvider.php
@@ -6,6 +6,7 @@ namespace Modules\Users\Providers;
 
 use Illuminate\Support\ServiceProvider;
 use Livewire\Livewire;
+use Modules\Users\Livewire\ProfileIndex;
 use Modules\Users\Livewire\UserIndex;
 
 class UsersServiceProvider extends ServiceProvider
@@ -19,5 +20,6 @@ class UsersServiceProvider extends ServiceProvider
     {
         $this->loadViewsFrom(__DIR__.'/../Resources/views', 'users');
         Livewire::component('users.index', UserIndex::class);
+        Livewire::component('profile.index', ProfileIndex::class);
     }
 }